Split out OpenSslError, ScopedSslBio, and trace macros
Also moved logging macros to macros.h since they are now needed by multiple files.
